+++ projects/ngx-scrollbar-demo/src/app/app.routes.ts
@@ -11,11 +11,7 @@ export const routes: Routes = [
     loadComponent: () => import('./material-page/material-page.component').then(c => c.MaterialPageComponent),
   },
   {
-    path: 'integration',
+    path: 'other-integrations',
     loadComponent: () => import('./integration-page/integration-page.component').then(c => c.IntegrationPageComponent),
-  },
-  // {
-  //   path: 'lab',
-  //   loadComponent: () => import('./lab/lab.component').then(c => c.LabComponent),
-  // }
+  }
 ];